+++ package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "devDependencies": {
+    "@types/node": "^24.0.10",
     "typescript": "~5.8.3",
     "vite": "^7.0.0"
   },
