+++ src/main.ts
@@ -1,42 +1,7 @@
 import "./style.css";
 
-import { unified } from "unified";
-import markdown from "remark-parse";
-import remark2rehype from "remark-rehype";
-import html from "rehype-stringify";
-
-const markdownToHtml = (markdownSource: string) => {
-  const html_text = unified()
-    .use(markdown)
-    .use(remark2rehype)
-    .use(html)
-    .processSync(markdownSource);
-
-  if (typeof html_text.value === "string") {
-    return html_text.value;
-  }
-
-  return "";
-};
-
-const blogList = () => {
-  const ul = document.createElement("ul");
-
-  const modules = import.meta.glob(`../content/**/**.md`, { eager: true });
-
-  for (const path in modules) {
-    const li = document.createElement("li");
-
-    const blogLink = document.createElement("a");
-    blogLink.innerText = path;
-    blogLink.href = path;
-    li.appendChild(blogLink);
-
-    ul.appendChild(li);
-  }
-
-  return ul;
-};
+import blogList from "./blogList";
+import markdownToHtml from "./markdownToHtml";
 
 const main = async () => {
   const readme = await import("../content/2025-07-02/2025-07-02-page.md?raw");
